requirements.txt
Django==5.2.11
djangorestframework
djangorestframework-simplejwt
django-filter